Read relationship command refactor
