--- public/DTO/IReadRelationshipByNamedDTO.tsx
@@ -1,6 +1,0 @@
-/**
- * A read relationship command using the 'named' argument expects to have a realtionship's name.
- */
-export default interface IReadRelationshipByNamedDTO {
-    relationshipName: string;
-}
