+++ public/DTO/IReadRelationshipDTO.tsx
@@ -1,0 +1,8 @@
+/**
+ * A read relationship command may accept two classes names or rallationship name.
+ */
+export default interface IReadRelationshipDTO {
+    relationshipName?: string;
+    sourceClassifierName?: string;
+    targetClassifierName?: string;
+}
